pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personalization-experimentation-platform"
version = "0.1.0"
description = "Reproducible offline feed personalization and experimentation pipelines."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.136.1",
  "httpx==0.28.1",
  "numpy==1.24.3",
  "pandas==1.5.3",
  "pyarrow==11.0.0",
  "pydantic==2.13.3",
  "pyyaml==6.0",
  "scikit-learn==1.3.0",
  "uvicorn==0.46.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit==4.6.0",
  "pytest==7.4.0",
]

[project.scripts]
personalization-show-blueprint = "personalization_platform.pipeline.show_blueprint:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
